# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample API to demonstrate Swagger with Gin.",
        "title": "ScheduleApiGo",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/consumer/start": {
            "post": {
                "description": "Publishes a job message to RabbitMQ with connection parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Consume"
                ],
                "summary": "Consume a job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "RabbitMQ host",
                        "name": "host",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "RabbitMQ port",
                        "name": "port",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "RabbitMQ user",
                        "name": "user",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "RabbitMQ password",
                        "name": "password",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "RabbitMQ queue",
                        "name": "queue",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Consumer success",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Error response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/job": {
            "get": {
                "description": "Retrieve a job by its ID from the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "job"
                ],
                "summary": "Get job by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Job"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new job to the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "job"
                ],
                "summary": "Create a new job",
                "parameters": [
                    {
                        "description": "Job data",
                        "name": "job",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Job"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/job/list": {
            "get": {
                "description": "Return a list of jobs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "job"
                ],
                "summary": "List of jobs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Job"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/server": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "server"
                ],
                "summary": "list of servers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Server"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new Server to database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "server"
                ],
                "summary": "Create a new server",
                "parameters": [
                    {
                        "description": "Object Data",
                        "name": "server",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Server"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Server"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Job": {
            "type": "object",
            "properties": {
                "cmdExecute": {
                    "type": "boolean",
                    "example": true
                },
                "date": {
                    "type": "string",
                    "example": "2025-02-04T15:04:05Z"
                },
                "description": {
                    "type": "string",
                    "example": "Daily backup job"
                },
                "hold": {
                    "type": "boolean",
                    "example": false
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Backup Job"
                },
                "priority": {
                    "type": "integer",
                    "example": 1
                },
                "script": {
                    "type": "string",
                    "example": "backup.sh"
                },
                "serverId": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "model.Server": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                }
            }
        }
    }
}